Reorganise Nunjucks global functions
